requirements.txt
Django==4.2.7
requests==2.31.0
python-dotenv==1.0.0
gspread==5.12.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
qrcode[pil]==7.4.2
Pillow==10.1.0
pytz==2023.3
